humanFriendlyStageKeyをutilに切り出す
